refactored Explore class; changed seed type in ReservoirSampling to long

+++ src/main/java/sampling/ReservoirSampler.java
@@ -18,7 +18,7 @@ public class ReservoirSampler {
      * Set the seed for the internal random generator. Used when one wants reproducibility.
      * @param seed: new seed
      */
-    public static void setSeed(int seed){
+    public static void setSeed(long seed){
         random.setSeed(seed);
     }
